# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: MetaRGB
    description: REST API for MetaRGB microservices. Laravel-compatible JSON responses served via Kong → grpc-gateway → gRPC services.
    title: MetaRGB API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: grpc-gateway direct (dev)
      url: http://localhost:8080
...
paths:
...
    /api/maps:
        get:
            operationId: get_maps
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: List maps
            tags:
                - Maps
    /api/maps/{id}:
        get:
            operationId: get_maps_id
            parameters:
                - description: Map ID
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: integer
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: Get map details
            tags:
                - Maps
...
components:
    schemas:
        Error:
            properties:
                message:
                    type: string
            type: object
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Laravel Sanctum-style bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http